requirements.txt
babel==2.18.0
blinker==1.9.0
flask-babel==4.0.0
cachelib==0.13.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
flask==3.1.2
flask-caching==2.3.1
flask-cors==6.0.2
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.3
markupsafe==3.0.3
pandas==3.0.2
pymysql==1.1.2
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.7.2
sqlalchemy==1.4.54
typing-extensions==4.15.0
urllib3==2.6.3
werkzeug==3.1.5
gunicorn==23.0.0
cryptography==46.0.7
coverage==7.13.5